# Simplify point-in-ring algorithm